Work towards more aggressive lexical lowering
This starts us down the path of turning "normal" variables into
locals, which will not only be a little faster for lookup, but give
dynamic optimizers a good bit more to work with too. It will also
enable more block flattening.

This gets the transformation essentially working, however there's
quite a bit of spectest fallout that needs to be investigated yet
(most likely with just a few root causes).